+++ src/lib.rs
@@ -429,6 +429,7 @@ pub mod pcm;
 pub mod pns;
 pub mod predictor;
 pub mod ps_data;
+pub mod ps_decoder;
 pub mod ps_decorr;
 pub mod ps_huffman;
 pub mod ps_hybrid;